+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.2
+
+- add new config for global disable
+- add new commands to disable/enable the auto selection globally & for the opened file